// sourcecode/BTrak.Packages/snova-trading-module/trading-module-components/trading-components/src/lib/billing/models/Contract-purchased.ts
export class ContractPurchaseModel {
    clientId: string
    contractFormId: string
    formJson: string
    formData: string
    fullName: string
    companyId: string
    searchText: string
    templateId: string;
    formName: string;
    purchaseName:string;
    purchaseId:string;
    isArchived: boolean;
    createdDateTime: any;
    createdBy: string;
    timeStamp: any;
}